backend/requirements.txt
asgiref==3.3.4
attrs==21.2.0
autobahn==21.3.1
automat==20.2.0
certifi==2021.5.30
cffi==1.14.5
channels==3.0.3
chardet==4.0.0
constantly==15.1.0
cryptography==3.4.7
daphne==3.0.2
defusedxml==0.7.1
django-allauth==0.44.0
django-cors-headers==3.7.0
django-rest-auth==0.9.5
django-rest-framework==0.1.0
django==3.2.4
djangorestframework==3.12.4
hyperlink==21.0.0
idna==2.10
incremental==21.3.0
oauthlib==3.1.1
pillow==8.2.0
psycopg2-binary==2.9.1
psycopg2==2.9.1
pyasn1-modules==0.2.8
pyasn1==0.4.8
pycparser==2.20
pyjwt==2.1.0
pyopenssl==20.0.1
python3-openid==3.2.0
pytz==2021.1
requests-oauthlib==1.3.0
requests==2.25.1
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.1
twisted==21.2.0
txaio==21.2.1
urllib3==1.26.5
zope.interface==5.4.0
